Use pointers from the caller for publishing status
This is so that the class instance can be used across multiple calls to
the run_node function. There's probably a cleaner way of doing this, but
the mbed way of doing stuff is pretty darn full of pitfalls... Ideally,
I want to have the Thread instance in the class, but apparently you
can't call the assignment operator for the Thread class.